Edit tgt macros

+++ data/static/tgt_consensus_a/macros/transform_file.sql
@@ -4,7 +4,7 @@ select
   filename::text as "filename",
   format::text as "format",
   data_type::text as "data_type",
-  size::integer as "size",
+  size::bigint as "size",
   drs_uri::text as "drs_uri",
   file_metadata::text as "file_metadata",
   has_access_policy::text as "has_access_policy",
